refactor: move petTypeColors file to public/data
