+++ public/data/petTypeColors.js
